backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-openai
pinecone
SQLAlchemy
psycopg2-binary
python-dotenv
tiktoken

